+++ src/main/java/lehjr/numina/common/capabilities/module/blockbreaking/IBlockBreakingModule.java
@@ -42,13 +42,13 @@ import net.minecraftforge.event.entity.player.PlayerEvent;
 import javax.annotation.Nonnull;
 
 public interface IBlockBreakingModule extends IPowerModule {
-    default boolean canHarvestBlock(@Nonnull ItemStack stack, BlockState state, Player player, BlockPos pos, int playerEnergy) {
+    default boolean canHarvestBlock(@Nonnull ItemStack stack, BlockState state, Player player, BlockPos pos, double playerEnergy) {
         if (playerEnergy >= this.getEnergyUsage() && isToolEffective(player.level, pos, getEmulatedTool()))
             return true;
         return false;
     }
 
-    boolean mineBlock(@Nonnull ItemStack powerFist, Level worldIn, BlockState state, BlockPos pos, LivingEntity entityLiving, int playerEnergy);
+    boolean mineBlock(@Nonnull ItemStack powerFist, Level worldIn, BlockState state, BlockPos pos, LivingEntity entityLiving, double playerEnergy);
 
     void handleBreakSpeed(PlayerEvent.BreakSpeed event);
