+++ src/main/java/lehjr/numina/common/config/ServerConfig.java
@@ -46,30 +46,30 @@ public class ServerConfig {
                 builder.push("Energy_Storage");
                 {
                         builder.push("battery_basic");
-                        builder.defineInRange("base_maxEnergy", 1.0E6D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxEnergy", 1.0E6D, 0, Integer.MAX_VALUE);
                         builder.define("isAllowed", true);
-                        builder.defineInRange("base_maxTransfer", 1.0E6D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxTransfer", 1.0E6D, 0, Integer.MAX_VALUE);
                         builder.pop();
                 }
                 {
                         builder.push("battery_advanced");
-                        builder.defineInRange("base_maxEnergy", 1.0E7D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxEnergy", 1.0E7D, 0, Integer.MAX_VALUE);
                         builder.define("isAllowed", true);
-                        builder.defineInRange("base_maxTransfer", 1.0E7D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxTransfer", 1.0E7D, 0, Integer.MAX_VALUE);
                         builder.pop();
                 }
                 {
                         builder.push("battery_elite");
-                        builder.defineInRange("base_maxEnergy", 1.0E8D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxEnergy", 1.0E8D, 0, Integer.MAX_VALUE);
                         builder.define("isAllowed", true);
-                        builder.defineInRange("base_maxTransfer", 1.0E8D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxTransfer", 1.0E8D, 0, Integer.MAX_VALUE);
                         builder.pop();
                 }
                 {
                         builder.push("battery_ultimate");
-                        builder.defineInRange("base_maxEnergy", 1.0E9D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxEnergy", 1.0E9D, 0, Integer.MAX_VALUE);
                         builder.define("isAllowed", true);
-                        builder.defineInRange("base_maxTransfer", 1.0E9D, 0, Double.MAX_VALUE);
+                        builder.defineInRange("base_maxTransfer", 1.0E9D, 0, Integer.MAX_VALUE);
                         builder.pop();
                 }
                 builder.pop(); // EnergyStorage
