+++ src/powersuits/java/lehjr/powersuits/common/item/module/miningenhancement/AdvancedVeinMiner.java
@@ -128,8 +128,8 @@ public class AdvancedVeinMiner extends AbstractPowerModule {
                     return false;
                 }
 
-                int playerEnergy = ElectricItemUtils.getPlayerEnergy(player);
-                int energyUsage = this.getEnergyUsage();
+                double playerEnergy = ElectricItemUtils.getPlayerEnergy(player);
+                double energyUsage = this.getEnergyUsage();
 
                 AtomicInteger bbModuleEnergyUsage = new AtomicInteger(0);
