+++ src/powersuits/java/lehjr/powersuits/common/item/module/tool/AxeModule.java
@@ -83,7 +83,7 @@ public class AxeModule extends AbstractPowerModule {
             }
 
             @Override
-            public boolean mineBlock(@NotNull ItemStack powerFist, Level worldIn, BlockState state, BlockPos pos, LivingEntity entityLiving, int playerEnergy) {
+            public boolean mineBlock(@NotNull ItemStack powerFist, Level worldIn, BlockState state, BlockPos pos, LivingEntity entityLiving, double playerEnergy) {
                 if (this.canHarvestBlock(powerFist, state, (Player) entityLiving, pos, playerEnergy)) {
                     ElectricItemUtils.drainPlayerEnergy((Player) entityLiving, getEnergyUsage());
                     return true;
